Inherit  from ApplicationRecord when on Rails 7
It appears that bindings have changed when on Rails 7 and inheriting from ActiveRecord::Base results in trying to make connection
through the activerecord adapter for File which fails with ActiveRecord::ConnectionNotEstablished. Given ApplicationRecord is suppose
to be the new parent class anyways, updating the logic accordingly.

Open to any feedback/suggestions here, thank you